signin, walklisten and play default

+++ src/imap/main/Huatong_dialog.java
@@ -1,9 +1,13 @@
 package imap.main;
 
+import imap.nettools.NetThread;
 import imap.nettools.Variable;
 
 import java.io.BufferedWriter;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.lang.reflect.Array;
@@ -24,9 +28,11 @@ import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.os.Environment;
+import android.view.KeyEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
@@ -69,18 +75,99 @@ public class Huatong_dialog extends Activity {
 		final String []check_sytle = {"未知","正式","狂野","欢乐","其他"};
 		
 		uploadButton.setOnClickListener(new View.OnClickListener() {
-			
+			//
 			@Override
 			public void onClick(View v) {
-				editText_name.getText().toString();
-				editText_brief.getText().toString();
-				// TODO 自动生成的方法存根
+				
+				String spotid = MainActivity.viewspotlist.get(view_point_sure_to_update).getId() + "";
+				String title = editText_name.getText().toString();
+				String description = editText_brief.getText().toString();
+				
+				RadioButton sexRadioButton = (RadioButton)findViewById(sexRadioGroup.getCheckedRadioButtonId());
+				RadioButton languageRadioButton = (RadioButton)findViewById(languageRadioGroup.getCheckedRadioButtonId());
+				RadioButton styleRadioButton = (RadioButton)findViewById(styleRadioGroup.getCheckedRadioButtonId());
+				
+				String style = null;
+		    	String language = null;
+		    	String gender = null;
+				for(int i = 0 ;i<5;i++)
+				{
+					if(sexRadioButton.getText().toString().equals(check_sex[i]))
+					{
+						style = i + "";
+					}
+				}
+				
+				for(int i = 0 ;i<5;i++)
+				{
+					if(languageRadioButton.getText().toString().equals(check_language[i]))
+					{
+						language = i + "";
+					}
+				}
+				for(int i = 0 ;i<5;i++)
+				{
+					if(styleRadioButton.getText().toString().equals(check_sytle[i]))
+					{
+						gender = i + "";
+					}
+				}
+				
 				File file_amr = new File(Variable.voicepath,
 						"voice_" + view_point_sure_to_update
 								+ "_" + timeString + ".amr");
-				if (file_amr.exists())
-					file_amr.delete();
-				finish();
+				
+				String fileaString= null;
+				try {
+					// 读取文件
+					FileInputStream fileInputStream = new FileInputStream(file_amr);
+					ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+					int i;
+					while ((i = fileInputStream.read()) != -1)
+						byteArrayOutputStream.write(i);
+					fileInputStream.close();
+					
+					// 把文件存在一个字节数组中
+					byte[] filea = byteArrayOutputStream.toByteArray();
+					byteArrayOutputStream.close();
+					fileaString = new String(filea, "ISO-8859-1");
+					
+					
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+				
+				SharedPreferences sp = Huatong_dialog.this.getSharedPreferences("imap", 0);
+				String username = sp.getString("username", "");
+				String password = sp.getString("password", "");
+				
+				NetThread netthread = new NetThread(username, password);
+				netthread.makeParam(Variable.sendVoice, spotid, fileaString, title, description, gender, language, style);
+				int returnCode = netthread.beginDeal();
+				
+				if (returnCode == 0)
+				{
+					if (file_amr.exists())
+						file_amr.delete();
+					finish();
+					
+					Toast.makeText(Huatong_dialog.this, "语音上传成功！请静待处理~", 
+				                 Toast.LENGTH_SHORT).show(); 
+					finish();
+				  }
+				  else if (returnCode == -1)
+				  {
+					  Toast.makeText(Huatong_dialog.this, "网络错误！", 
+				                 Toast.LENGTH_SHORT).show();
+				  }
+				  else
+				  {
+					  Toast.makeText(Huatong_dialog.this, Variable.errorCode[returnCode] + "！", 
+				                 Toast.LENGTH_SHORT).show();
+				  }
+				
+				
+				
 			}
 		});
 
@@ -89,16 +176,8 @@ public class Huatong_dialog extends Activity {
 			@Override
 			public void onClick(View v) {
 				//标签文件地址
-				File file_txt = new File(Variable.voicepath,
-						"voice_" + view_point_sure_to_update
-								+ "_" + timeString + ".txt");
+				
 				try {
-					BufferedWriter writer = new BufferedWriter(
-							new FileWriter(file_txt));
-					//添加选择框的值
-					RadioButton sexRadioButton = (RadioButton)findViewById(sexRadioGroup.getCheckedRadioButtonId());
-					RadioButton languageRadioButton = (RadioButton)findViewById(languageRadioGroup.getCheckedRadioButtonId());
-					RadioButton styleRadioButton = (RadioButton)findViewById(styleRadioGroup.getCheckedRadioButtonId());
 					
 					if(editText_name.getText().toString().isEmpty())
 					{
@@ -119,6 +198,18 @@ public class Huatong_dialog extends Activity {
 				                 Toast.LENGTH_SHORT).show(); 
 						return;
 					}
+					
+					File file_txt = new File(Variable.voicepath,
+							"voice_" + view_point_sure_to_update
+									+ "_" + timeString + ".txt");
+					
+					BufferedWriter writer = new BufferedWriter(
+							new FileWriter(file_txt));
+					//添加选择框的值
+					RadioButton sexRadioButton = (RadioButton)findViewById(sexRadioGroup.getCheckedRadioButtonId());
+					RadioButton languageRadioButton = (RadioButton)findViewById(languageRadioGroup.getCheckedRadioButtonId());
+					RadioButton styleRadioButton = (RadioButton)findViewById(styleRadioGroup.getCheckedRadioButtonId());
+					
 					writer.write(editText_name.getText().toString() + "\n");
 					writer.write(editText_brief.getText().toString()+ "\n");
 					
@@ -184,5 +275,11 @@ public class Huatong_dialog extends Activity {
 		
 		
 	}
+	
+	@Override 
+	 public boolean onKeyUp(int keyCode, KeyEvent event) 
+	 {
+		return true; 
+	} 
 }
